Move ReactTestFiberComponent functions from Renderer to Component file
